requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
azure-core==1.27.1
azure-storage-blob==12.16.0
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.1
Django==4.2.2
gunicorn==20.1.0
idna==3.4
isodate==0.6.1
kombu==5.3.1
prompt-toolkit==3.0.38
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==4.5.4
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.5.0
